Provide Retrofit and IApi via DI module rather than instantiating it in ApiClient.
